requirements.txt
appnope==0.1.2
asgiref==3.3.1
attrs==20.3.0
backcall==0.2.0
decorator==4.4.2
Django==3.1.5
django-debug-toolbar==3.2
django-rest-framework==0.1.0
djangorestframework==3.12.2
iniconfig==1.1.1
ipdb==0.13.4
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
packaging==20.8
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.16
ptyprocess==0.7.0
py==1.10.0
Pygments==2.7.4
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
python-decouple==3.4
pytz==2020.5
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
